Merged PR 202275: Replace SAM with SAM-HQ and extend the `generate_masks` to work with batch of images
The aim of this PR is to substitute the SAM model with the SAM-HQ model. To get the best performance out of SAM-HQ, images are first arranged in a square grid structure with some padding. This results in a better segmentation mask and much faster mask generation (around 15x). The final segmentation maks are unpacked to get the mask for each individual image before saving 'em.