Merged PR 202461: Pill segmentation modularization with SAM model
The aim of this PR is to provide the modularization of pill segmentation using online SAM models. It links the segmentation model with the image object in _image.py_. The _segmenter_example.py_ file could be run in _python segmenter_example.py_. Also, _requirements.txt_ file is updated for adding new libraries. A _config.yml_ file is also added to include the parameters that the pill segmentation needs.