Merged PR 202726: Extend `RxImageGenerator` to work for remote server
The aim of this PR is to extend the `RxImageGenerator` class to work for images and masks stored in our remote directory. It also modifies the `fetch_data` function to raise a `Timeout` error when a connection couldn't be established to the remote server. Additionally, the dataset generator README is also updated.