Merged PR 207205: Change `types` name to `pipelines` for `mypy` to work for all the scripts
The `types` folder name conflicted with the built-in types, resulting in skipping all the files in the `types` folder. This is the error we used to get when running the linters in Azure Pipeline.
```
mypy: "types/__init__.py" shadows library module "types"
note: A user-defined top-level module with the name "types" is not supported
```
Hence, the folder name has been changed to `pipelines` for `mypy` to run successfully.

Also, a bug was fixed in the Azure pipeline that resulted in linters passing even when there were some issues with linters. Additionally, some missing type-stubs were also added.

`mypy` and `pre-commit` has been updated to the latest version.