Merged PR 216232: Port Segmentation model using Pytorch Trial API
This PR incorporates the PyTorch Trial API for the Semantic Segmentation model. The old dataset, models, and training scripts have been moved to the `archive` folder. Also, some of the packages that are required for the training have been moved to the `requirements-training` text file. More importantly, the `torch` version has been downgraded to the `1.12` version for the training to work properly on the cluster.