Merged PR 219741: Create `RxSemanticSegmentation` object for model inference
Add the `RxSemanticSegmentation` object in the pipelines for segmenting pills from the background. Additionally, a device argument is added in the `RxBase` to control the device on which to place the model. The `__call__` method is implemented as well. The types in the `BasicAugTransform` are enhanced for both images and masks. The `SegmentTransform` is simplified slightly.

Cherry picked from !218880

Cherry-picked from commit `37709e07`.