Merged PR 220412: Ability to control the `Color` and `Shape` of the Pills generated
Extend the data generation pipeline to control the pill's colors and shapes to generate synthetic images.
In order to generate images with specific types of colors and/or shapes, one needs to pass the `--colors` and/or `--shapes` flag to the `dataset_generate` script. For instance, to generate only the RED and BLACk color and CAPSULE shape pills, one can do it using:
```
dataset_generate --colors Black --colors Red --shape Capsule
```
Some other routine improvements are made with respect to the new features added.