Merged PR 224040: Add multiple sampling strategies inside the dataset generator pipeline
A `--sampler` flag is added that can be used to control the behavior of the way pills are sampled. Additionally, the `--colors` and `--shapes` flags are modified so that the list of colors and/or shapes can be passed as a single string. For instance, to use red and white color pills, one can do that using `--colors "red, white"`.